+++ CHANGELOG.md
@@ -17,6 +17,8 @@ Test and beta releases will have experimental functions enabled by default.
 **Next release**
 
 * Removed restriction *getPackagesForUid*
+* Run migration in background and added message *migration complete*
+* Attempt to fix security exception
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)