+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Test and beta releases will have experimental functions enabled by default.
 * Attempt to fix security exception
 * Moved meta.xml to Meta class
 * Cache permissions
+* Optimized usage data getting (solves performance problems)
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)