+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Test and beta releases will have experimental functions enabled by default.
 * Moved meta.xml to Meta class
 * Cache permissions for better performance
 * Optimized usage data getting (should solve most of the performance issue)
+* Half check box for restricted IPC category (special case)
 * Use inbox style notifications ([issue](https://github.com/M66B/XPrivacy/issues/1112))
 
 **Version 1.99.1 EXPERIMENTAL**