+++ CHANGELOG.md
@@ -16,13 +16,15 @@ Test and beta releases will have experimental functions enabled by default.
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 1.99.3 EXPERIMENTAL**
+
 * Fixed security exception reported with support info
 * Fixed restart required notification
 * Option to register device on custom e-mail address
-* User defined dangerous functions, long press a function name to toggle ([issue](https://github.com/M66B/XPrivacy/issues/869))
 * Restored restrictions *getNetwork...* and *getSim...* ([issue](https://github.com/M66B/XPrivacy/issues/1125))
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+* User defined dangerous functions, long press a function name to toggle ([issue](https://github.com/M66B/XPrivacy/issues/869))
 
 **Version 1.99.2 EXPERIMENTAL**